service-backend-a/pyproject.toml
[tool.poetry]
authors = ["Tomonori Hayashi"]
description = "FastAPI OpenTelemetry Sample"
name = "src"
readme = "README.md"
version = "0.2.0"

[tool.poetry.dependencies]
aiosqlite = ">=0.19"
fastapi = {extras = ["all"], version = ">=0.104"}
greenlet = "^3.0"
python = "^3.11"
sqlalchemy = "^2.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.24.0"
opentelemetry-exporter-gcp-trace = "^1.6.0"
opentelemetry-instrumentation-sqlite3 = "^0.45b0"
opentelemetry-instrumentation-sqlalchemy = "^0.45b0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = ">=0.21"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
ignore_missing_imports = true
